added the ai backend
