+++ flask/.env.template
@@ -2,3 +2,7 @@ MONGO_URI=
 BOOTSTRAP_SERVERS=
 SASL_USERNAME=
 SASL_PASSWORD=
+ANTHROPIC_API_KEY=
+VOYAGE_API_KEY=
+HF_TOKEN=
+OPENAI_API_KEY=
