+++ flask/misinformationCategories.txt
