--- flask/parsing/.env
@@ -1,3 +1,0 @@
-ANTHROPIC_API_KEY=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
-VOYAGE_API_KEY=pa-tnbMCP7aN1Qsf6by6QNVbCSQWjT0MACDMeqamxtcLi9
-HF_TOKEN=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
